requirements.txt
python-craigslist>=1.1.4
sqlalchemy
python-dateutil
slack_sdk>=3.15.2
